+++ lib/rley/syntax/grm_symbol.rb
@@ -2,13 +2,16 @@ module Rley # This module is used as a namespace
   module Syntax # This module is used as a namespace
     # Abstract class for grammar symbols.
     # A grammar symbol is an element that appears in grammar rules.
-    class GrmSymbol
+    class GrmSymbol  
       # The name of the grammar symbol
       attr_reader(:name)
 
+      # Constructor.
+      # aName [String] The name of the grammar symbol.
       def initialize(aName)
         @name = aName.dup
       end
+        
     end # class
   end # module
 end # module